move test file into trainer test explicitly
